[feat #142] prompt들 파일로 따로 뺴서 import로 사용하기
