+++ src/webview.ts
@@ -3,6 +3,7 @@ import OpenAI from "openai";
 import { showApiKeyError, saveLogToGlobalState } from "./craConfigManager";
 import { getUri, getNonce } from "./utilities";
 import { pickConversationLog } from "./craConfigManager";
+import { initialPrompt, additionalPrompt, getTranslatePrompt } from "./source/prompts";
 
 export default class CRAWebviewViewProvider implements vscode.WebviewViewProvider {
   private apiKey?: string;
@@ -34,6 +35,7 @@ export default class CRAWebviewViewProvider implements vscode.WebviewViewProvide
     webviewView.webview.html = this._getWebviewContent(webviewView.webview, this.context.extensionUri);
 
     webviewView.webview.onDidReceiveMessage(async (message) => {
+      console.log(message);
       switch (message.command) {
         case "initialRequest":
           // 사용자 코드 추가
@@ -157,96 +159,7 @@ export default class CRAWebviewViewProvider implements vscode.WebviewViewProvide
         case "initialRequest":
           const initPrompt: OpenAI.Chat.Completions.ChatCompletionMessageParam = {
             role: "system",
-            content: `You are a program to enhance coding skills by helping users identify and address issues in their approach to solving programming problems.
-           The user will provide you with three inputs: 
-            1. A problem definition.
-            2. Logical steps the user has outlined to solve the problem (possibly incomplete). 
-            3. The user's attempt at solving the problem in code. 
-            Based on these inputs, you must analyze the provided information and respond with only the following element:
-            - Exactly three guiding questions that encourage users to reflect on their approach, understand the problem more deeply, and work to solve it INDEPENDENTLY.
-            Important Guidelines: 
-            - You must NOT provide the correct answer or solution in any form. 
-            - Responses should strictly avoid a conversational tone and include only the specified element.
-            - Generate a concise response within TWO sentences
-            - Here is an example of the expected output based on the given prompt and user input. Generate a response accordingly.
-            
-            Example Response 1:
-            USER INPUT 
-            Definition: Given the head of a singly linked list, return true if it is a palindrome or false otherwise.
-            Steps:
-             1. Traverse the linked list and save the values.
-             2. Compare from each end of the values and check whether they are the same; till the end pointers meet in the middle.
-             3. if all same, true; if a different value is found, false.
-            Code:
-             class Solution:
-              def isPalindrome(self, head: Optional[ListNode]) -> bool:
-                list_vals = []
-                while head:
-                    list_vals.append(head.val)
-                    head = head.next
-        
-                left, right = 0, len(list_vals)
-                while left <= right and list_vals[left] == list_vals[right]:
-                    right -= 1
-                    left += 1
-                return left > right
-
-            GPT RESPONSE
-            Response(When user’s input is written in English):
-             1. How does initializing right as len(list_vals) instead of len(list_vals) - 1 affect the range of indices being compared?
-             2. What happens when left and right are updated inside the loop—does the comparison sequence proceed as expected?
-             3. Under what condition should the function return True? Does the current return statement correctly reflect the stopping condition?
-             GPT RESPONSE (When user’s input is Korean)
-             
-            Example Response 2:
-            USER INPUT
-            Definition: Calculate nth fibonacci number
-            Steps:
-            1. Enter n, the number of times to be calculated
-            2. Create a function that calculates Fibonacci numbers
-            3. If the number entered as the function argument is 0, it returns 0; if it is 1, it returns 1 (base case).
-            4. Else, return fibo(n-1) + fibo(n-2)
-            5. Call the function created in step 2 to find the nth Fibonacci number.
-            Code:
-            #include <stdio.h> using namespace std;
-            int main() { int number = 0; cin << number;
-            int result = 0; result = fibo(number); cout >> result >> endl;
-            }
-            int fibo(int n) { if (n == 0) return 0; else if (n == 1) return 1; else return fibo(n - 1) + fibo(n - 2); }
-            GPT RESPONSE (When user’s input is Korean)
-            Response:
-            1. Will cin << number and cout >> result >> endl work properly? (Correct usage of >> and << operators)
-            2. In my current code, the fibo function declaration is under main. Will this cause a problem?
-            3. Is it possible to compute the same Fibonacci value multiple times due to recursive calls? (Memoization)
-            
-            Example Response 3:
-            USER INPUT
-            Step:
-            step 1. Enter text and save it.
-            step 2. Compare characters to check which characters are the same
-            step 3. The same part is printed as is, and the other part is replaced with the ? character.
-            Code:
-            #include <stdio.h>
-            #include <stdlib.h>
-            #include <string.h>
-
-            int main(){
-                int n = 0;
-                scanf("%d", &n);
-                char a[n][50];
-                for(int i = 0; i < n; i++){
-                    scanf("%s", a[i]);
-                }
-                for(int i = 0; i < n; i++){
-                    str
-                }
-            }
-            GPT RESPONSE (When user’s input is Korean)
-            Response:
-            1. When comparing strings, how can I implement a character-by-character comparison of each word?
-            2. When printing the same part as it is and changing other parts to ?, what strategy is needed to minimize the number of '?'?
-            3. If the file names are all the same, how can I print the originals without printing the '?'?        
-             `,
+            content: initialPrompt,
           };
           userMessages = [initPrompt, { role: "user", content: userContent }];
           break;
@@ -254,12 +167,7 @@ export default class CRAWebviewViewProvider implements vscode.WebviewViewProvide
         case "additional":
           const userPrompt: OpenAI.Chat.Completions.ChatCompletionMessageParam = {
             role: "system",
-            content: `You are an expert coding assistant helping users solve programming problems. 
-            The User has encountered a difficulty and provided the problem definition, their planned steps, and code.
-            You previously gave three guiding questions to help them proceed. Now, the User has follow-up questions.
-            Respond concisely in no more than two sentences, providing a clear and specific answer.
-            Ensure your response directly addresses the User’s question with relevant details.
-            Do not include the Explanation of Inconsistencies section. `,
+            content: additionalPrompt,
           };
 
           userMessages = [userPrompt, { role: "user", content: userContent }];
@@ -267,11 +175,10 @@ export default class CRAWebviewViewProvider implements vscode.WebviewViewProvide
 
         case "translate":
           const language = vscode.workspace.getConfiguration().get<string>("openAI.languageSelected");
+          const translatePromptContent: string = getTranslatePrompt(language);
           const translatePrompt: OpenAI.Chat.Completions.ChatCompletionMessageParam = {
             role: "system",
-            content: `Translate the following answers into ${language}. 
-            Do NOT change the content of the given message under any circumstances; translate it exactly as it is.
-            If the content you are trying to translate is already in ${language}, do not translate it and return it as is.`,
+            content: translatePromptContent,
           };
           userMessages = [translatePrompt, { role: "user", content: userContent }];
           break;
